+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 12.0.0-beta.7 (Unreleased)
 
+### New Features
+
+- Added support for `SetProperties` in share client. This API supports update share tier and adjusting share's quota.
+- Added support to get share's tier status in `ListSharesSinglePage` and `GetProperties`.
+
 ### Breaking Changes
 
 - Removed `GetDirectoryClient` and `GetFileClient` from `ShareClient`. `ShareDirectoryClient` and `ShareFileClient` now initializes with the name of the resource, not path, to indicate that no path parsing is done for the API
@@ -18,6 +23,7 @@
 - Removed `c_` for constants: `c_FileDefaultTimeValue`, `c_FileCopySourceTime`, `c_FileInheritPermission`, `FilePreserveSmbProperties` and `FileAllHandles`.
 - `Concurrency`, `ChunkSize` and `InitialChunkSize` were moved into `DownloadShareFileToOptions::TansferOptions`.
 - `Concurrency`, `ChunkSize` and `SingleUploadThreshold` were moved into `UploadShareFileFromOptions::TransferOptions`.
+- Removed `SetQuota` related API, result and options. The functionality is moved into `SetProperties`.
 
 ### Other Changes and Improvements